# src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Entity Browser Service",
    "version" : "1.0.0",
    "title" : "laaws-template",
    "contact" : {
      "name" : "LOCKSS Support",
      "url" : "http://www.lockss.org/",
      "email" : "support@lockss.org"
    },
    "license" : {
      "name" : "BSD 3-Clause License",
      "url" : "https://www.tldrlegal.com/l/bsd3"
    }
  },
  "host" : "127.0.0.1:8888",
  "basePath" : "/tmpl",
  "tags" : [ {
    "name" : "Default",
    "description" : "Default section"
  }, {
    "name" : "entities",
    "description" : "tasks to browse entities"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/entities/all" : {
      "get" : {
        "tags" : [ "entity" ],
        "summary" : "Return all entities",
        "description" : "Returns all entities in the collection",
        "operationId" : "getAllEntities",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Return the list of entities",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/entity"
              }
            }
          }
        }
      }
    },
    "/entities/one" : {
      "get" : {
        "tags" : [ "entity" ],
        "summary" : "Return one entity",
        "description" : "Returns one entity at random",
        "operationId" : "getOneRandomEntity",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Returns one entity at random",
            "schema" : {
              "$ref" : "#/definitions/entity"
            }
          }
        }
      }
    },
    "/entities/{id}" : {
      "get" : {
        "tags" : [ "entity" ],
        "summary" : "Return an entity with id",
        "description" : "Returns one entity with the passed id",
        "operationId" : "getOneEntityById",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "An entities id",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Return the entity with the specified id",
            "schema" : {
              "$ref" : "#/definitions/entity"
            }
          },
          "404" : {
            "description" : "No such entity exists"
          }
        }
      }
    }
  },
  "definitions" : {
    "entity" : {
      "type" : "object",
      "required" : [ "id", "name", "value" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The id assigned to this entity"
        },
        "name" : {
          "type" : "string",
          "description" : "The name of this entity"
        },
        "value" : {
          "type" : "string",
          "description" : "The value of this entity"
        }
      },
      "description" : "A simple entity"
    }
  }
}